+++ autd3-firmware-emulator/tests/op/debug.rs
@@ -26,7 +26,7 @@ fn send_debug_output_idx(
     let mut cpu = CPUEmulator::new(0, geometry.num_transducers());
     let mut tx = vec![TxMessage::new_zeroed(); 1];
 
-    let d = DebugSettings::new(|_, gpio| match gpio {
+    let d = GPIOOutputs::new(|_, gpio| match gpio {
         GPIOOut::O0 => debug_types[0].clone(),
         GPIOOut::O1 => debug_types[1].clone(),
         GPIOOut::O2 => debug_types[2].clone(),
@@ -47,7 +47,7 @@ fn send_debug_pwm_out() -> anyhow::Result<()> {
     let mut cpu = CPUEmulator::new(0, geometry.num_transducers());
     let mut tx = vec![TxMessage::new_zeroed(); 1];
 
-    let d = DebugSettings::new(|dev, gpio| match gpio {
+    let d = GPIOOutputs::new(|dev, gpio| match gpio {
         GPIOOut::O0 => DebugType::PwmOut(&dev[0]),
         GPIOOut::O1 => DebugType::PwmOut(&dev[1]),
         GPIOOut::O2 => DebugType::PwmOut(&dev[2]),